requirements.txt
customtkinter
moviepy
Pillow